+++ Source/radiation/Radiation.cpp
@@ -647,12 +647,6 @@ Radiation::Radiation(Amr* Parent, Castro* castro, int restart)
     std::cout << "temp_floor    = " << temp_floor << std::endl;
   }
 
-  if (do_real_eos == 0) {
-      if (const_c_v <= 0.0) {
-          amrex::Abort("do_real_eos == 0 requires const_c_v > 0");
-      }
-  }
-
   if (verbose > 2) {
     Vector<int> temp;
     if (pp.queryarr("spot",temp,0,BL_SPACEDIM)) {